Drop Object
Place the Object_Drop script into the Game Object where object want to be dropped. Make sure the Game Object has a slot and assign the slot into the Drop Slot column. If the Game Object scale isn't 1,1,1, make empty Game Object and use it as the Game Object Parent